+++ wssAgileEngine/src/main/java/com/mycompany/wssAgileEngine/model/Transaction.java
@@ -10,20 +10,20 @@ public class Transaction {
 
     private String id;
     private String type;
-    private Float amount;
+    private Double amount;
     private Calendar effectiveDate;
 
     public Transaction() {
     }
 
-    public Transaction(String id, String type, Float amount, Calendar effectiveDate) {
+    public Transaction(String id, String type, Double amount, Calendar effectiveDate) {
         this.id = id;
         this.type = type;
         this.amount = amount;
         this.effectiveDate = effectiveDate;
     }
 
-    public Transaction(String type, Float amount, Calendar effectiveDate) {
+    public Transaction(String type, Double amount, Calendar effectiveDate) {
         this.type = type;
         this.amount = amount;
         this.effectiveDate = effectiveDate;
@@ -45,11 +45,11 @@ public class Transaction {
         this.type = type;
     }
 
-    public Float getAmount() {
+    public Double getAmount() {
         return amount;
     }
 
-    public void setAmount(Float amount) {
+    public void setAmount(Double amount) {
         this.amount = amount;
     }
